backend/pyproject.toml
[project]
name = "crowdforge"
version = "0.1.0"
description = "CrowdForge backend API"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1,<6.0",
    "django-ninja>=1.3,<2.0",
    "psycopg[binary]>=3.2,<4.0",
    "django-redis>=5.4,<6.0",
    "redis>=5.0,<6.0",
    "meilisearch>=0.31,<1.0",
    "pydantic>=2.9,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.9",
    "factory-boy>=3.3",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "crowdforge.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
pythonpath = [".", "backend"]

[tool.ruff]
target-version = "py312"
line-length = 120

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
